Codec.Serialise.IO

Description

High-level file-based API for serialising and deserialising values.

FilePath API

writeFileSerialise

Arguments

:: Serialise a

=> FilePath
The file to write to.

-> a
The value to be serialised and written.

-> IO ()

Serialise a ByteString and write it directly to the specified file.

Since: 0.2.0.0

readFileDeserialise

Arguments

:: Serialise a

=> FilePath
The file to read from.

-> IO a
The deserialised value.

Read the specified file (internally, by reading a ByteString) and attempt to decode it into a Haskell value using deserialise (the type of which is determined by the choice of the result type).

Throws: DeserialiseFailure if the file fails to deserialise properly.

Since: 0.2.0.0

Handle API

hPutSerialise

Arguments

:: Serialise a

=> Handle
The Handle to write to.

-> a
The value to be serialised and written.

-> IO ()

Serialise a ByteString (via serialise) and write it directly to the specified Handle.

Since: 0.2.0.0