Codec.Serialise.Internal.GeneralisedUTF8

Documentation

encodeGenUTF8 :: String -> (SlicedByteArray, UTF8Encoding)

Encode a string as (generalized) UTF-8. In addition to the encoding, we return a flag indicating whether the encoded string contained any surrogate characters, in which case the output is generalized UTF-8.

data UTF8Encoding

Constructors

ConformantUTF8

GeneralisedUTF8

decodeGenUTF8 :: ByteArray -> String

Utilities

isSurrogate :: Char -> Bool

Is a Char a UTF-16 surrogate?

isValid :: UTF8Encoding -> [Word8] -> Bool

Is the given byte sequence valid under the given encoding?